Erreur pour récupérer le catalogue de formation dès lors qu'un élève est lié a une formation. Pas d'erreur avant ça.

+++ Spring/ITraining_API/src/main/java/com/example/itraining_api/repository/LearnerAccountRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.itraining_api.entity.LearnerAccount;
 
-public interface LearnerAccountRepository extends CrudRepository<LearnerAccount, Long> {
+public interface LearnerAccountRepository extends CrudRepository<LearnerAccount, Integer> {
 
 }
